BotSimulation/requirements.txt
pygame==2.1.0
pyserial==3.5
